% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_define_haplotypes.R
\name{collate_define_haplotypes}
\alias{collate_define_haplotypes}
\title{Collate define_haplotypes Lists}
\usage{
collate_define_haplotypes(haplotype_list)
}
\arguments{
\item{haplotype_list}{A list of the lists created by the define_haplotypes function.}
}
\value{
A collated list of all haplotype tables.
}
\description{
This function collates a list of output files from define_haplotypes.
}
\examples{
collate_define_haplotypes(list)
}
