% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_haplotype_variants.R
\name{collate_haplotype_variants}
\alias{collate_haplotype_variants}
\title{Collate haplotype_variants Tables}
\usage{
collate_haplotype_variants(haplotype_variants_list, format = 1)
}
\arguments{
\item{haplotype_variants_list}{A list of the tables created by the define_haplotypes function.}

\item{format}{The format you want the output table to be in. This should be the same number you used when running define_haplotypes.}
}
\value{
A collated table of haplotype variants.
}
\description{
This function collates a list of output files from haplotype_variants.
}
\examples{
collate_haplotype_variants(list, format = 1)
}
